+++ src/olmo_core/optim/__init__.py
@@ -1,11 +1,18 @@
 from .adamw import AdamWConfig
 from .config import OptimConfig, OptimGroupOverride
+from .lion import Lion, LionConfig, SkipStepLion, SkipStepLionConfig
 from .scheduler import ConstantScheduler, CosWithWarmup, Scheduler
+from .skip_step_optimizer import SkipStepOptimizer
 
 __all__ = [
     "OptimConfig",
     "OptimGroupOverride",
+    "SkipStepOptimizer",
     "AdamWConfig",
+    "LionConfig",
+    "Lion",
+    "SkipStepLionConfig",
+    "SkipStepLion",
     "Scheduler",
     "CosWithWarmup",
     "ConstantScheduler",